pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "kafka-forwarder-configurer"  # REQUIRED, is the only field that cannot be marked as dynamic.
dynamic = ["version"]
description = "ISIS Kafka forwarder configurer."
readme = "README.md"
requires-python = ">=3.12"
license-files = ["LICENSE"]

authors = [
  {name = "ISIS Experiment Controls", email = "ISISExperimentControls@stfc.ac.uk" }
]
maintainers = [
  {name = "ISIS Experiment Controls", email = "ISISExperimentControls@stfc.ac.uk" }
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
"pyepics",
"ess-streaming-data-types",
"genie-python",
"ibex-non-ca-helpers",
  "kafka-python"
]

[project.optional-dependencies]
doc = [
  "sphinx", 
  "sphinx_rtd_theme", 
  "myst_parser",
  "sphinx-autobuild",
  "sphinx-codeautolink",
  "sphinxcontrib-mermaid",
  "sphinxcontrib-spelling",
]
lint = [
  "pyright",
  "ruff",
]
test = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
]
dev = [
  "kafka-forwarder-configurer[doc,lint,test]",
]

[project.urls]
"Homepage" = "https://github.com/isiscomputinggroup/kafka_forwarder_configurer"
"Bug Reports" = "https://github.com/isiscomputinggroup/kafka_forwarder_configurer/issues"
"Source" = "https://github.com/isiscomputinggroup/kafka_forwarder_configurer"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"
addopts = "--cov --cov-report=html -vv"
filterwarnings = [
    'ignore::tango.PyTangoUserWarning',
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
  "version.py",  # Autogenerated
]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if typing.TYPE_CHECKING:",
  "@abstractmethod",
]

[tool.coverage.html]
directory = "coverage_html_report"

[tool.pyright]
include = ["src", "tests"]
reportConstantRedefinition = true
reportDeprecated = true
reportInconsistentConstructor = true
reportMissingParameterType = true
reportMissingTypeArgument = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnnecessaryIsInstance = true
reportUntypedBaseClass = true
reportUntypedClassDecorator = true
reportUntypedFunctionDecorator = true

#[tool.setuptools_scm]
#version_file = "src/kafka_forwarder_configurer/version.py"

[tool.build_sphinx]
